added common base class for Avro record writers
